backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
pgvector>=0.2.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Face Recognition
face-recognition>=1.3.0
numpy>=1.26.0
pillow>=10.0.0

# Utils
python-dotenv>=1.0.0
setuptools>=70.0.0  # Required for face_recognition_models in Python 3.12+

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
